consolidate PivotSyncState
